+++ source/sha512.cpp
@@ -339,7 +339,7 @@ namespace CrissCross
 			Reset();
 			if (!_length || !_data) return -1;
 
-			sha512_update(&m_state, (const unsigned char *)_data, _length);
+			sha512_update(&m_state, (const unsigned char *)_data, (unsigned int)_length);
 			m_hash = new unsigned char[SHA512_DIGEST_SIZE];
 			sha512_final(&m_state, m_hash);
 			return 0;
@@ -372,7 +372,7 @@ namespace CrissCross
 
 			if (!_data) return -1;
 
-			sha512_update(&m_state, (unsigned char *)_data, _length);
+			sha512_update(&m_state, (unsigned char *)_data, (unsigned int)_length);
 			return 0;
 		}
